requirements.txt
fastapi
uvicorn
python-dotenv
requests
pypdf
faiss-cpu
numpy
sentence-transformers
python-multipart
streamlit